TES3::Cell changes and docs.
- Add docs for TES3::Cell, TES3::CellExteriorData, and TES3::PackedColor.
- Change helper functions to use float for water level.
- Prevent getting region from interiors.

+++ MWSE/TES3Cell.h
@@ -80,8 +80,8 @@ namespace TES3 {
 		int getFogDensity();
 		void setFogDensity(int);
 
-		int getWaterLevel();
-		void setWaterLevel(int);
+		float getWaterLevel();
+		void setWaterLevel(float);
 
 		Region * getRegion();
 	};
